requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
pymongo==4.6.1
python-socketio==5.11.2
psutil==5.9.8
numpy==1.26.3
joblib==1.3.2
watchdog==4.0.0
gunicorn==22.0.0